things

+++ src/ch/zombieInvasion/util/Util.java
@@ -4,20 +4,21 @@ import java.util.Random;
 
 
 public class Util {
-	static Random rnd = new Random();
+  static Random rnd = new Random();
 
-	/**
-	 * Returns a float between -1 and 1
-	 * 
-	 * @return float
-	 */
-	public static float smallRndFloat() {
-		return rnd.nextFloat() - rnd.nextFloat();
-	}
+  /**
+   * Returns a float between -1 and 1
+   * 
+   * @return float
+   */
+  public static float smallRndFloat() {
+    return rnd.nextFloat() - rnd.nextFloat();
+  }
 
-	public static final float map(float value, float istart, float istop, float ostart, float ostop) {
-		return ostart + (ostop - ostart) * ((value - istart) / (istop - istart));
-	}
+  public static final double map(double value, double istart, double istop, double ostart,
+      double ostop) {
+    return ostart + (ostop - ostart) * ((value - istart) / (istop - istart));
+  }
 
   public static String uniqueID() {
     return java.util.UUID.randomUUID().toString();
